Fix scores
以下修正
・scoresのデータ取得
・scoresのグラフスタイル
・firebaseのデプロイ
・ロゴ追加
