+++ .firebase/hosting.ZGlzdA.cache
@@ -1,0 +1,9 @@
+favicon.ico,1625481121959,1e71457865f706dc865b49a54a86e193818220d290b30226b6630a42faf1535d
+index.html,1625481121959,12c4f43c2bd0f892a2d783ecc43f43cbed00ecfd44d63e17236664b9c25903ee
+css/app.23c15894.css,1625481121998,f6e6a3d3c6baf1e73d9fb6638056314f323437ea129f9d0fd8e775b754dd7409
+css/chunk-vendors.ca7f275c.css,1625481122004,6128824d51f8a9f9d325b7ee0c5aa4e93d6733d65470c802ddf0338db64a5533
+img/Logo.f47e63e5.png,1625481121984,76d9d520abc102bd47e98af6cb373416d7a9a198e06d0e3c821305fee4738aeb
+js/app.67918d93.js,1625481122001,fb7a0ef5df87571ab0ad5dfe5a6aa227e2efc6a2ea802c6e58e3cf173370af0c
+js/app.67918d93.js.map,1625481122003,bdb54acd1a3542001bafc49569a4c4f8cc86a3b9f2c063ba3e9e11cb9a4f52f2
+js/chunk-vendors.02ae5425.js,1625481122003,c0fa4e67223b26eede548f5766d2e63aecf40832ac4d83fccafce16d1a60b9e8
+js/chunk-vendors.02ae5425.js.map,1625481122003,20fda00c767bd6a5890e6c3167b08572118a51524ff79af4b98b0499acdb04a8
