+++ .firebaserc
@@ -1,0 +1,5 @@
+{
+  "projects": {
+    "default": "lightq-e5c74"
+  }
+}
